nanomax-installer (6.0.max4)   * do-install.sh: Copy kernel images if exists
git-svn-id: http://max.educa.madrid.org:8000/svn/trunk@996 424cea3e-0240-0410-9575-8bd47b58cd33

+++ nanomax-installer/usr/share/nanomax-installer/do-install.sh
@@ -100,7 +100,7 @@ copiar_nanomax() {
   # mover el kernel
   echo " * Configurando el kernel..."
   for f in initrd.gz minirt.gz vmback vmlinuz; do
-    mv /mnt/nanomax/casper/$f /mnt/nanomax/
+    [ -e /mnt/nanomax/casper/$f ] && mv /mnt/nanomax/casper/$f /mnt/nanomax/
   done
 }
